requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Graph Processing
networkx>=3.1
neo4j>=5.12.0
pyvis>=0.3.2

# Vector DB & RAG
chromadb>=0.4.15
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10

# NLP & Information Extraction
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
markdown>=3.5
pytesseract>=0.3.10
Pillow>=10.0.0

# Video/Audio Transcription
openai-whisper>=20230918
youtube-transcript-api>=0.6.1
yt-dlp>=2023.10.13

# API & Web
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
flask>=3.0.0
flask-cors>=4.0.0

# UI
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0
click>=8.1.0
watchdog>=3.0.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
seaborn>=0.12.0
matplotlib>=3.7.0

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
ipykernel>=6.26.0
jupyter>=1.0.0
